requirements.txt
# Image Summarizer Requirements
# Core dependencies for web server and basic functionality

# Configuration management
pyyaml>=6.0.0

# Web server dependencies
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# Image processing
pillow>=10.0.0

# AI/ML dependencies for local models (LLaVA, Falconsai)
# Default configuration uses local models - these are required
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Optional cloud provider dependencies
# Uncomment as needed:

# For Amazon Bedrock
# boto3>=1.28.0
# botocore>=1.29.0

# For Azure OpenAI and OpenAI
# openai>=1.0.0

# Development and testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0